+++ package.json
@@ -29,7 +29,7 @@
     "axios": "^0.27.2",
     "crypto-js": "^4.1.1",
     "next": "12.2.5",
-    "next-auth": "4.18.8",
+    "next-auth": "^4.22.1",
     "nodemailer": "^6.9.1",
     "nuka-carousel": "^5.4.0",
     "postcss-nesting": "^10.2.0",
@@ -44,7 +44,7 @@
     "recoil": "^0.7.6",
     "recoil-persist": "^4.2.0",
     "sharp": "^0.32.0",
-    "sweetalert2": "^11.7.3",
+    "sweetalert2": "^11.4.8",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
